// src/generated/types.ts
    export type id = number;
// ...
export namespace AssetFields {
    export type id = number;
    export type account = string;
    export type assetId = string;
    export type name = string;
    export type supply = string;
    export type precision = number;
    export type symbol = string;
    export type txHash = string;

}

export interface Asset {
    id: AssetFields.id;
    account: AssetFields.account;
    assetId: AssetFields.assetId;
    name: AssetFields.name;
    supply: AssetFields.supply;
    precision: AssetFields.precision;
    symbol: AssetFields.symbol;
    txHash: AssetFields.txHash;

}